+++ src/jpeg.cpp
@@ -4,14 +4,14 @@
 
 #include "../include/jpeg.hpp"
 
-void get_jpeg_dimensions(std::istream &file, size_t *height, size_t *width) {
+void get_jpeg_dimensions(std::istream &file, uint *height, uint *width) {
 
 }
 
-bool read_jpeg(std::istream &file, uint32_t **pixels, size_t *height_ptr, size_t *width_ptr) {
+bool read_jpeg(std::istream &file, uint32_t **pixels, uint *height_ptr, uint *width_ptr) {
     return true;
 }
 
-bool save_jpeg(std::ostream &file, uint32_t **pixels, size_t height, size_t width, void *args) {
+bool save_jpeg(std::ostream &file, uint32_t **pixels, uint height, uint width, void *args) {
     return true;
 }
