+++ src/ppm.cpp
@@ -5,7 +5,7 @@
 #include "../include/ppm.hpp"
 #include "../include/utils.hpp"
 
-void get_ppm_dimensions(std::istream &file, size_t *height, size_t *width) {
+void get_ppm_dimensions(std::istream &file, uint *height, uint *width) {
     std::string format {};
     uint8_t buffer;
 
@@ -13,13 +13,13 @@ void get_ppm_dimensions(std::istream &file, size_t *height, size_t *width) {
     file >> format >> *width >> *height >> buffer >> buffer >> buffer;
 }
 
-bool read_ppm(std::istream &file, uint32_t **pixels, size_t *height_ptr, size_t *width_ptr) {
+bool read_ppm(std::istream &file, uint32_t **pixels, uint *height_ptr, uint *width_ptr) {
     uint8_t b {}, g {}, r {};
     
     get_ppm_dimensions(file, height_ptr, width_ptr);
 
-    for (size_t  y = 0; y < *height_ptr; y++)
-        for (size_t x = 0; x < *width_ptr; x++) {
+    for (uint  y = 0; y < *height_ptr; y++)
+        for (uint x = 0; x < *width_ptr; x++) {
             file >> r >> g >> b;
             pixels[y][x] = get_hex(r, g, b);
         }
@@ -27,12 +27,12 @@ bool read_ppm(std::istream &file, uint32_t **pixels, size_t *height_ptr, size_t 
     return true;
 }
 
-bool save_ppm(std::ostream &file, uint32_t **pixels, size_t height, size_t width, void *args) {
+bool save_ppm(std::ostream &file, uint32_t **pixels, uint height, uint width, void *args) {
     file << "P6\n" << width << " " << height << "\n255\n";
     uint8_t r {}, g {}, b {};
 
-    for (size_t  y = 0; y < height; y++)
-        for (size_t x = 0; x < width; x++) {
+    for (uint  y = 0; y < height; y++)
+        for (uint x = 0; x < width; x++) {
             get_rgb(pixels[y][x], &r, &g, &b);
             file << r << g << b;
         }
